Add onGenerateRoute for routing and update OnboardingView
Implemented the `onGenerateRoute` function to handle routing in the application. Added a static `routeName` to `OnboardingView` for easier navigation. Updated the `OnboardingView` to use `AppStyles` for text styling and adjusted the image properties for better display. Modified `main.dart` to use the new routing function and set the initial route to `OnboardingView`.